Fix: Notifications not sending when using default setting (#183)
When users try to send notifications but they have not yet set their preference in the user settings menu, the notifier selected was None. This fix defines what the default notifier is and defaults to this when sending messages to other users or yourself when your preference is not set.