Ensure s3 chunk read accounts for utf8 length (#296)
The s3 read is chunked by some number of bytes. Since utf8 characters are dynamically lengthed, we would have a s3 chunk that starts and ends with an invalid utf8 character. The issue is resolved by always breaking a s3 chunk into a valid part and an invalid part, and then merge the invalid part to the next chunk.