feat: Allow disabling Query Engines
This feature allows Query Engines to be disabled without deleting or removing them from an environment.  Disabled Query Engines cannot be used to execute queries and they do not appear in the Query Engine dropdown.  They do appear correctly where used in DataDocs and snippets, and appear in the search filters.